+++ R/zzz.R
@@ -1,0 +1,6 @@
+#' Environment to log unique entries related with downloading problems
+#' @return Empty env
+#' @keywords internal
+#' @export
+#' @noRd
+env <- new.env(parent = emptyenv())